requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.0
    # via requests
deprecated==1.2.15
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-semantic-conventions
googleapis-common-protos==1.66.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.68.1
    # via opentelemetry-exporter-otlp-proto-grpc
idna==3.10
    # via requests
importlib-metadata==8.5.0
    # via opentelemetry-api
opentelemetry-api==1.29.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.29.0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-common==1.29.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.29.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.29.0
    # via opentelemetry-exporter-otlp
opentelemetry-proto==1.29.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.29.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.50b0
    # via opentelemetry-sdk
protobuf==5.29.1
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
requests==2.32.3
    # via opentelemetry-exporter-otlp-proto-http
typing-extensions==4.12.2
    # via opentelemetry-sdk
urllib3==2.2.3
    # via requests
wrapt==1.17.0
    # via deprecated
zipp==3.21.0
    # via importlib-metadata
